src/font_generation/requirements.txt
anyio==3.6.2
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
idna==3.3
markdown-it-py==2.1.0
mdurl==0.1.2
msgpack==1.0.3
nala==0.12.2
olefile==0.46
pexpect==4.8.0
Pillow==9.4.0
ptyprocess==0.7.0
pycurl==7.45.2
Pygments==2.14.0
pynvim==0.4.2
pyparsing==3.0.9
PySimpleSOAP==1.16.2
python-apt==2.6.0
python-debian==0.1.49
python-debianbts==4.0.1
reportbug==12.0.0
requests==2.28.1
rfc3986==1.5.0
rich==13.3.1
six==1.16.0
sniffio==1.2.0
socksio==1.0.0
tomli==2.0.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
